make calculate_long_adjustment use decimal

+++ elfpy/markets/hyperdrive/hyperdrive_actions.py
@@ -501,8 +501,8 @@ def calculate_short_adjustment(
 def calculate_long_adjustment(
     market_state: hyperdrive_market.MarketState,
     position_duration: time.StretchedTime,
-    market_time: float,
-) -> float:
+    market_time: Decimal,
+) -> Decimal:
     """Calculates an adjustment amount for lp shares"""
     if market_time > market_state.long_average_maturity_time:
         return 0
